+++ HISTORY.md
@@ -311,7 +311,7 @@
 
 0.0.15 (2024/11/12 18:04:29)
 ------------------------------
-- [DictAttrAnnotRequired] add  
+- [DictGaAnnotRequired] add  
 - [_SAND] create and add first content  
 
 0.0.14 (2024/11/08 17:51:33)
@@ -351,7 +351,7 @@
 
 0.0.5 (2024/11/05 10:33:43)
 ------------------------------
-- [classes] add DictIgnorecase+DictAttr  
+- [classes] add DictIcKeys+DictGa  
 - [stock] move here  
 - [all] fix almost, so tests works (need fix privates for alerts)